197.691: two cruxes with the word "abounded."
CRUX 1: On pg 197 / XML ll. 691, 7 lines up from the bottom on the EEBO scan, the passage "Byam was one that abounded in his own wit" the reads as follows on the XML: 

abound<gap reason="illegible" extent="1 letter">
                  <desc>•</desc>
               </gap>d

Which, when I did a trial run encoding to HTML, came out as "abound[.]d." The previous coder apparently wasn't confident that this is an "e": it's extremely faint apart from the curve at the bottom, which may have as much to do with the quality of EEBO's photocopy as with any damaged type. But comparing the e's to the periods in Oroonoko, I think it's safe to say this is an e. 

CRUX 2. Interestingly, where Todd has "abounded his own wit" (pg. 65; in her notes she parses as "Was secure in his own opinion"), our text has "abounded in his own Wit," which is closer to the idiom noted in the OED and to my ears emphasizes Byam's hubris a bit more. Since Todd is also using the first edition of the novel as the basis of her text, the discrepancy suggests there may be variant states of the edition on that page. 

MV